BMI and Gender Prediction/requirements.txt
streamlit
opencv-python
keras
tensorflow
keras_vggface
cvlib
numpy
Pillow
